utils: move define_device from main
